terragrunt/aws/glue/curated/platform/gc_notify/requirements.txt
awswrangler==3.12.0
pandas==2.2.3
python-dateutil==2.9.0
